Fix delayed call of plot `_on_close()`
