+++ hyperspy/drawing/image.py
@@ -266,12 +266,6 @@ class ImagePlot(BlittedFigure):
 
         self.connect()
 
-    def add_marker(self, marker):
-        marker.ax = self.ax
-        if marker.axes_manager is None:
-            marker.axes_manager = self.axes_manager
-        self.ax_markers.append(marker)
-
     def update(self, auto_contrast=None, **kwargs):
         ims = self.ax.images
         # Turn on centre_colormap if a diverging colormap is used.
@@ -410,13 +404,6 @@ class ImagePlot(BlittedFigure):
             optimize_for_oom(step_oom - i)
             i += 1
 
-    def _on_close(self):
-        for marker in self.ax_markers:
-            marker.close()
-        self.events.closed.trigger(obj=self)
-        for f in self.events.closed.connected:
-            self.events.closed.disconnect(f)
-        self.figure = None
-
     def close(self):
-        plt.close(self.figure)  # This will trigger self._on_close()
+        self._on_close()   # Needs to trigger serially for a well defined state
+        plt.close(self.figure)
